+++ web/src/features/project/AdminPlaceholder.tsx
@@ -1,5 +1,6 @@
 import { AdminPage } from "./AdminPage";
+import { getCurrentDateString } from "./DashboardPage";
 
 export function AdminPlaceholder() {
-  return <AdminPage />;
+  return <AdminPage today={getCurrentDateString()} />;
 }